+++ src/domain/sim-time-converter.ts
@@ -1,3 +1,11 @@
+/*
+NOTE: convertDuration and convertRate should not be widely used. Prefer convertComponent and convertLevel instead.
+
+Primitives for converting between real-world values and values that humans
+can perceive i.e. a latency of 5ms is a flicker, but 500ms can be observed.
+
+*/
+
 /** Ratio of visual animation time to real-world time. A component with a
  * real-world latency of 10ms will animate for 10 * TIME_SCALE = 1000ms. */
 const TIME_SCALE = 100;